bug fixes and factoring out code
