+++ calyx-py/test/correctness/queues/rr_queues/rr_queue_7flows.data
@@ -50,7 +50,52 @@
       0,
       2,
       2,
-      2
+      2,
+      0,
+      0,
+      2,
+      0,
+      1,
+      1,
+      0,
+      1,
+      1,
+      0,
+      0,
+      1,
+      0,
+      1,
+      1,
+      2,
+      2,
+      0,
+      2,
+      2,
+      2,
+      1,
+      0,
+      1,
+      1,
+      1,
+      1,
+      2,
+      1,
+      0,
+      1,
+      1,
+      2,
+      0,
+      0,
+      1,
+      0,
+      2,
+      1,
+      1,
+      0,
+      2,
+      1,
+      1,
+      1
     ],
     "format": {
       "is_signed": false,
@@ -60,56 +105,101 @@
   },
   "values": {
     "data": [
-      104,
-      93,
-      353,
-      100,
-      196,
-      152,
-      11,
-      184,
-      212,
-      84,
-      74,
-      135,
-      33,
-      169,
-      154,
-      308,
-      300,
-      1,
-      305,
-      347,
-      362,
-      173,
-      33,
-      158,
-      181,
-      156,
-      246,
-      356,
-      161,
-      94,
-      246,
-      241,
-      360,
-      90,
-      29,
-      131,
-      11,
-      383,
-      183,
-      206,
-      9,
-      281,
-      214,
-      187,
-      192,
       296,
       4,
       231,
       23,
-      362
+      362,
+      92,
+      319,
+      100,
+      60,
+      386,
+      125,
+      236,
+      176,
+      262,
+      181,
+      268,
+      128,
+      397,
+      236,
+      55,
+      301,
+      383,
+      399,
+      188,
+      151,
+      18,
+      221,
+      46,
+      106,
+      174,
+      262,
+      312,
+      185,
+      75,
+      174,
+      141,
+      359,
+      279,
+      47,
+      159,
+      351,
+      162,
+      156,
+      90,
+      40,
+      320,
+      76,
+      369,
+      352,
+      158,
+      247,
+      82,
+      368,
+      24,
+      41,
+      307,
+      273,
+      207,
+      16,
+      121,
+      379,
+      304,
+      176,
+      128,
+      233,
+      333,
+      215,
+      74,
+      28,
+      326,
+      16,
+      252,
+      171,
+      106,
+      66,
+      374,
+      288,
+      67,
+      322,
+      211,
+      54,
+      86,
+      222,
+      190,
+      76,
+      30,
+      215,
+      150,
+      72,
+      232,
+      317,
+      86,
+      267,
+      232,
+      249
     ],
     "format": {
       "is_signed": false,
@@ -119,6 +209,51 @@
   },
   "ans_mem": {
     "data": [
+      0,
+      0,
+      0,
+      0,
+      0,
+      0,
+      0,
+      0,
+      0,
+      0,
+      0,
+      0,
+      0,
+      0,
+      0,
+      0,
+      0,
+      0,
+      0,
+      0,
+      0,
+      0,
+      0,
+      0,
+      0,
+      0,
+      0,
+      0,
+      0,
+      0,
+      0,
+      0,
+      0,
+      0,
+      0,
+      0,
+      0,
+      0,
+      0,
+      0,
+      0,
+      0,
+      0,
+      0,
+      0,
       0,
       0,
       0,
